+++ contracts/interfaces/revest/IRewardsHandler.sol
@@ -3,21 +3,20 @@
 pragma solidity >=0.8.0;
 
 interface IRewardsHandler {
-
     struct UserBalance {
-        uint allocPoint; // Allocation points
-        uint lastMul;
+        uint256 allocPoint; // Allocation points
+        uint256 lastMul;
     }
 
-    function updateShares(uint fnftId, uint newShares) external;
+    function updateShares(uint256 fnftId, uint256 newShares) external;
 
-    function getAllocPoint(uint fnftId) external view returns (uint);
+    function getAllocPoint(uint256 fnftId) external view returns (uint256);
 
-    function claimRewards(uint fnftId, address caller) external returns (bool);
+    function claimRewards(uint256 fnftId, address caller) external returns (bool);
 
     function setStakingContract(address stake) external;
 
-    function getRewards(uint fnftId) external view returns (uint);
+    function getRewards(uint256 fnftId) external view returns (uint256);
 
-    function depositReward(uint amount) external;
+    function depositReward(uint256 amount) external;
 }
